PG-311: Fix user/system cpu time zero values.

+++ pg_stat_monitor.c
@@ -1359,8 +1359,8 @@ pgss_update_entry(pgssEntry *entry,
             }
 		    if (sys_info)
 		    {
-			    e->counters.sysinfo.utime += (int64)(sys_info->utime - e->counters.sysinfo.utime)/e->counters.calls.calls;
-			    e->counters.sysinfo.stime += (int64)(sys_info->stime - e->counters.sysinfo.stime)/e->counters.calls.calls;
+			    e->counters.sysinfo.utime += (sys_info->utime - e->counters.sysinfo.utime)/e->counters.calls.calls;
+			    e->counters.sysinfo.stime += (sys_info->stime - e->counters.sysinfo.stime)/e->counters.calls.calls;
 		    }
 		    if (walusage)
 		    {
